requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
pypdf2==3.0.1
pdfplumber==0.10.3
pdfminer.six==20221105
pymupdf==1.23.8
aiofiles==23.2.1
jinja2==3.1.2
google-generativeai>=0.8.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart==0.0.6
requests>=2.31.0
apscheduler>=3.10.4
openai>=1.0.0

